requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gradio==4.22.0
gradio_client==0.13.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest_asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
opencv-python==4.9.0.80
orjson==3.9.15
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.1.0
prompt-toolkit==3.0.42
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
pydicom==2.1.0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-gdcm==3.0.23
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.4
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
semantic-version==2.10.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.36.3
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
torchxrayvision==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
typer==0.10.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
wheel==0.41.2
zipp==3.17.0
